requirements.txt
spotipy==2.22.0
requests==2.28.1
oauth2client==4.1.3
youtube-title-parse==1.0.0
google-api-python-client==2.71.0
python-dotenv==0.21.0